extend icon switching to main menu
