+++ src/App.tsx
@@ -4,7 +4,7 @@ import Container from 'react-bootstrap/Container';
 import Nav from 'react-bootstrap/Nav';
 import { useState, useEffect } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faStar as fasStar, faFilter, faBook, faHeart, faCheck, faMagnifyingGlass, faCalendarDays, faComment } from '@fortawesome/free-solid-svg-icons';
+import { faStar as fasStar, faFilter, faBook, faHeart, faCheck, faMagnifyingGlass, faCalendarDays, faComment, IconDefinition, faToriiGate, faBroom } from '@fortawesome/free-solid-svg-icons';
 import { faGithub } from '@fortawesome/free-brands-svg-icons';
 import FilterOptions from "./FilterOptions"
 import MenuPage from "./MenuPage"
@@ -14,6 +14,9 @@ import NavDropdown from 'react-bootstrap/NavDropdown';
 import Offcanvas from 'react-bootstrap/Offcanvas';
 import { Image } from 'react-bootstrap';
 import DarkModeSelector from './DarkModeSelector';
+import { COLORSTATUS } from './Utils';
+import { Icon } from '@fortawesome/fontawesome-svg-core';
+
 const touhoufest = require("./images/touhoufest.jpg");
 
 function App({ menupagedata, menuheader }) {
@@ -39,6 +42,24 @@ function App({ menupagedata, menuheader }) {
   // const [activeDayIndex, setActiveDayIndex] = useState(0);
   // const [summonDayScroll, setSummonDayScoll] = useState(0);
 
+  // state variables/functions for setting color theme
+  // (i cant beievei its this much typescript all for changing the color theme :skull:)
+  const [oppositecolorState, setOppositeColorState] = useState(getColorState());
+  function getColorState() {
+      let status:string|null = localStorage.getItem(COLORSTATUS);
+      if(status === "light" || status === null) {
+          document.documentElement.setAttribute('data-bs-theme','light');
+          return faBroom;
+      }
+      else {
+          document.documentElement.setAttribute('data-bs-theme','dark');
+          return faToriiGate;
+      }
+  }
+  function grabTrueColorState(input:IconDefinition) {
+    if(input === faToriiGate) {return faBroom;}
+    else{ return faToriiGate;}
+  }
 
   function changeMenuPageState(idx, isDisplayed) {
     let newstate = [...menupagebools];
@@ -167,7 +188,7 @@ function App({ menupagedata, menuheader }) {
             >
               <Offcanvas.Header closeButton>
                 <Offcanvas.Title id={`offcanvasNavbarLabel-expand-false`}>
-                  {menuheader}
+                  <span><FontAwesomeIcon icon={grabTrueColorState(oppositecolorState)}/> {menuheader}</span>
                 </Offcanvas.Title>
               </Offcanvas.Header>
               <Offcanvas.Body className="p-0">
@@ -175,7 +196,7 @@ function App({ menupagedata, menuheader }) {
                 <div className="p-3">
                   <Nav className="justify-content-end flex-grow-1 pe-3">
                     {menunavs}
-                    <Nav.Link><DarkModeSelector /></Nav.Link>
+                    <Nav.Link><DarkModeSelector oppositecolorState={oppositecolorState} setOppositeColorState={setOppositeColorState}/></Nav.Link>
                     <Nav.Link href="https://google.com" target="_blank"><FontAwesomeIcon icon={faComment} fixedWidth/> Feedback Form</Nav.Link>
                     <Nav.Link href="https://github.com/kir12/touhoufest-webapp" target="_blank"><FontAwesomeIcon icon={faGithub} fixedWidth></FontAwesomeIcon> About App</Nav.Link>
                     <Nav.Link href="https://www.google.com/search?q=marisa+kirisame&client=firefox-b-1-d&source=lnms&tbm=isch&sa=X&ved=2ahUKEwioqcvz4fT9AhW2kYkEHTCND3AQ0pQJegQIBBAC&biw=1920&bih=884&dpr=1" target="_blank"><FontAwesomeIcon icon={faHeart} fixedWidth></FontAwesomeIcon> Best Girl</Nav.Link>
